requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
einops
transformers
setuptools
ont-bonito
mamba-ssm
causal-conv1d
flash-attn --no-build-isolation # added later and didn't test